Eea phase2 (#955)
* add eea_createPrivacyGroup

* add eea_deletePrivacyGroup

* add eea_getPrivacyPrecompileAddress

* add eea_getPrivateTransaction

* Add support to for privacy groups

Encoder and Decoder support for:
* privateFrom privateFor
* privateFrom privacyGroupId
* privacyGroupId only

* Complete eea_getPrivateTransaction

* remove zombie code

* eea_findPrivacyGroup

* improve code coverage

* use getters in equals and hashcode methods
* remove setters for response objects

* checkstyle

* complete integration

* fix tests

* checkstyle

* move PantheonQuickstartIntegrationTest to integration-test package
* delete redundant generated contract

* clarify comment

* Make class field final

* version bump

* fix retrieving address with big chain id

* spotless after rebase

* Move priv methods to Pantheon module

* Fix Pantheon integration test

* Fix tests

* Fix tests

* Inject gas provider into PrivateTransactionManager

* Expose legacy privacy group generation

* wip: extended testing for priv_getPrivateTransaction

* make privateFrom mandatory

* remove tests/constructs that exclude privateFrom

* fix test

* Refactor and make base64 string type safe

* Fix integration test

* add tests for utils

* Address (most) comments

* Remove unused methods in Numeric
* Delegate to Enum.valueOf in Restriction
* Throw IllegalArgumentException in Base64String
* Use JsonDeserialize and JsonSerialize for Base64String
* Remove checkstyle related comments

* Suppress javadoc warning

+++ crypto/src/main/java/org/web3j/crypto/SignedRawTransaction.java
@@ -35,11 +35,11 @@ public class SignedRawTransaction extends RawTransaction implements SignatureDat
     }
 
     @Override
-    public byte[] getEncodedTransaction(Integer chainId) {
+    public byte[] getEncodedTransaction(Long chainId) {
         if (null == chainId) {
             return TransactionEncoder.encode(this);
         } else {
-            return TransactionEncoder.encode(this, chainId.byteValue());
+            return TransactionEncoder.encode(this, chainId);
         }
     }
 }
